+++ CHANGELOG.md
@@ -1,9 +1,20 @@
-# Change Log
+# EasyTask Change Log
 
 All notable changes to the "easytask" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.3.0] - 2024-02-11
+### Changed
+- HTML structure
+- CSS Styles
+- Refactoring
+- Icon
+
+### Added:
+- Regex to detect single and block comments in different languages.
+- Refresh button.
+- License.
+
+## [0.0.1] - 2024-02-10
 
 - Initial release